Fix bug with convoke where colored mana could end up being used as colorless.
Before this patch, the code would try to pay the mana symbols as they appear - for example if you convoke with a white creature and click "white" in the dialog, but the cost for the spell is "1WW", it would actually use that white mana to pay for the "1" portion of the cost, leaving "WW" to be paid (which is problematic if you have a swamp and a plains untapped which you were planning to use in addition to the white creature).

This change moves the logic for paying the cost from the UI code to the engine code (and re-uses existing code there) and also adds some unit tests for it.

git-svn-id: http://svn.slightlymagic.net/forge/trunk@28332 269b9781-a132-4a9b-9d4e-f004f1b56b58