Fix shard sort order when deciding if a mana cost is payable.
Prior to this change, the logic would conclude that a {1}{W} cost is not payable with an untapped Scoured Barrens (w/b) and Jungle Hollow (b/g) on the field.

The code is unfortunately too tangled (has too many dependencies) to write a unit test for this. Ideally, we should refactor so it wouldn't depend on the player, etc and just take parameters that can be fed test data.

git-svn-id: http://svn.slightlymagic.net/forge/trunk@28472 269b9781-a132-4a9b-9d4e-f004f1b56b58